Fixed issue resulting in segmentation fault with unsupported input.
When a WGFile was created with NTWPR_WGR_convertSU and the number
of nodes specified was 0. It allowed any edge number from the input file
to the WGFile. That was bad because some files like web-Google.txt contained
edges outside the [0, nodeNum-1] radius resulting in segmentation fault when
the WGFile was loaded in RAM with NTWPR_WGF_load2crs() function.